llmops-api/requirements.txt
openai
langchain-openai
langchain-community
langchain-core
python-dotenv
flask-migrate
Flask==3.1.2
injector==0.22.0
